# doc/api-specification.yml
openapi: 3.0.0
info:
  description: TARA login service
  version: 1.0.0-oas3
  title: TARA login service
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
  /auth/init:
    get:
      tags:
        - oidc-server-integration
      summary: Start the login session
      operationId: init
      description: |
        Verifies the authentication request.
        Initializes the authentication flow.
      parameters:
        - in: query
          name: login_challenge
          description: The login challenge from the OIDC service to initialize login session
          required: true
          schema:
            type: string
            pattern: 'a-zA-Z0-9'
            minimum: 0
            maximum: 50
            example: '824113da6da44889943b98715a06d699'
        - in: query
          name: lang
          description: Language for UI initialization
          schema:
            type: string
            enum: [et, en, ru]
      responses:
        '200':
          description: HTML response with a list of eligible login methods
          headers:
            Set-Cookie:
              description: SESSION=abcde12345; Path=/; HttpOnly; Secure
              schema:
                type: string
          content:
            text/html:
              schema:
                type: string
                example: 'The HTML login view is displayed'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/html:
              schema:
                type: string
                example: 'The HTML error view is displayed'
        '500':
          description: An internal error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            text/html:
              schema:
                type: string
                example: 'The HTML error view is displayed'
...
servers:
  - url: 'https://localhost:8443'
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: The time the error occurred
          example: '2020-12-22T09:09:23.463+00:00'
        status:
          type: string
          description: The HTTP statuscode
          example: '400'
        error:
          type: string
          description: Short error message
          example: 'Bad Request'
        message:
          type: string
          description: Human readable error description
          example: Required String parameter 'sample_parameter' is not present
        path:
          type: string
          description: The path that returned the error
          example: '/sample/path'
        locale:
          type: string
          description: The locale from the user session (if present)
          example: 'en_US'
        incident_number:
          type: string
          description: Unique ID to identify the error in logs
          example: 98JAE6AL9VGY2OFQ
...